Stagging

+++ public/app/src/jenis_kegiatan.js
@@ -28,6 +28,7 @@ var data_jenis_kegiatan = (function () {
                 { data: "jenis_kegiatan" },
                 { data: "short_id" },
                 { data: "code_id" },
+                {},
                 { data: "created_at" },
                 { data: "updated_at" },
                 {},
@@ -37,6 +38,18 @@ var data_jenis_kegiatan = (function () {
                     targets: 0,
                     searchable: false,
                     orderable: false,
+                },
+                {
+                    targets: -4,
+                    searchable: true,
+                    orderable: true,
+                    render: function (data, type, full, meta) {
+                        if (full.deleted_at === null) {
+                            return "Aktif";
+                        } else {
+                            return "Tidak Aktif";
+                        }
+                    },
                 },
                 {
                     targets: -3,
@@ -67,20 +80,37 @@ var data_jenis_kegiatan = (function () {
                     orderable: false,
                     render: function (data, type, full, meta) {
                         var editRoute = route + "/" + full.id + "/edit";
-
-                        return (
+                        if (full.deleted_at === null) {
+                            return (
+                                `
+                           <a href="` +
+                                editRoute +
+                                `" class="btn btn-sm btn-clean btn-icon btn-icon-md" title="Ubah">
+                              <i class="fa fa-pencil"></i>
+                            </a>
+                            <a data-id=` +
+                                full.id +
+                                ` href="#" onclick="deleteJenisKegiatan(this,event)" class="btn btn-sm btn-clean btn-icon btn-icon-md" title="Hapus">
+                              <i class="fa fa-trash"></i>
+                            </a>
                             `
-                       <a href="` +
-                            editRoute +
-                            `" class="btn btn-sm btn-clean btn-icon btn-icon-md" title="Ubah">
-                          <i class="fa fa-pencil"></i>
-                        </a>
-                        <a data-id=` +
-                            full.id +
-                            ` href="#" onclick="deleteJenisKegiatan(this,event)" class="btn btn-sm btn-clean btn-icon btn-icon-md" title="Hapus">
-                          <i class="fa fa-trash"></i>
-                        </a>`
-                        );
+                            );
+                        } else {
+                            return (
+                                `
+                                <a href="` +
+                                editRoute +
+                                `" class="btn btn-sm btn-clean btn-icon btn-icon-md" title="Ubah">
+                                  <i class="fa fa-pencil"></i>
+                                </a>
+                                <a data-id=` +
+                                full.id +
+                                ` href="#" onclick="restoreJenisKegiatan(this,event)" class="btn btn-sm btn-clean btn-icon btn-icon-md" title="Restore">
+                                  <i class="fa fa-mail-reply"></i>
+                                </a>
+                                `
+                            );
+                        }
                     },
                 },
             ],
@@ -98,6 +128,28 @@ var data_jenis_kegiatan = (function () {
 jQuery(document).ready(function () {
     data_jenis_kegiatan.init();
 });
+
+window.restoreJenisKegiatan = (input) => {
+    var restoreRoute = route + "/" + $(input).attr("data-id") + "/restore";
+    Swal.fire({
+        title: "Konfirmasi",
+        text: "Anda yakin ?",
+        icon: "warning",
+        showCancelButton: true,
+        confirmButtonColor: "#3085d6",
+        cancelButtonColor: "#d33",
+        confirmButtonText: "Ya, Restore",
+        cancelButtonText: "Tidak",
+        reverseButtons: false,
+    }).then((result) => {
+        if (result.value) {
+            updateData(restoreRoute).then((res) => {
+                Swal.fire("Sukses", "Data berhasil diaktifkan", "success");
+                window.location.reload();
+            });
+        }
+    });
+};
 
 window.deleteJenisKegiatan = (input) => {
     var deleteRoute = route + "/" + $(input).attr("data-id");
